moving Button into component
