+++ src/components/Button/Button.module.scss
@@ -1,4 +1,4 @@
-@import "../styles/globals.module.scss";
+@import "../../styles/globals.module.scss";
 
 .bouncing {
   @include bouncingO;
